refactor(install): lift mechanical install core into src/core/install
`daft install`, `daft repo install`, and `daft clone --install` all share the
same mechanical bootstrap — write the starter daft.yml, propagate it across a
repo's worktrees, and offer the visitor-privacy `.git/info/exclude` entry — but
that core lived inside the `commands::install` vertical slice, so `clone` and
`repo install` reached across command modules to call it.

Move the *how* into `src/core/install/` (the starter template, `install_at`,
`install_starter`, `propagate_starter_to_worktrees`, `InstallOptions`, and the
git-exclude cluster) alongside its 10 unit tests. The *where/whether* — argv
parsing, repo-position dispatch, and existing-config guidance — stays in
`commands::install` with its 6 dispatch tests. `clone` and `repo install` now
depend on the core module instead of a sibling command.

`install_starter` narrows from `pub` to `pub(crate)` (it is a crate-internal
building block, never part of the command surface); every other moved body is
byte-identical. starter.yml moves via `git mv`.
