+++ src/core/install/starter.yml
